Silvers
# Description
This adds the ability to serialize draw commands to .sriv file via a SerializingFactory which can be used by the unit tests in place of the NoOpFactory.

# Details
You can use the SerializingFactory to initialize your .riv and annotate the serialization with things like frame size and new frame/frame begin. This makes it possible to drop the generated .sriv into the Flutter viewer which will perform the serialized rendering commands to show exactly what the runtime is doing, high level rendering-wise. It can also diff two .sriv files if they are dropped in one drag/drop op.

When adding new "Silver" tests make sure to rebaseline the silvers by doing:
```
cd rive/packages/runtime/tests/unit_tests
./test.sh rebaseline
```

You can now modify the runtime and re-run the tests to make sure the silvers still match:
```
./test.sh
```

The important function to call at the end of your test is ::matches on the SerializingFactory. This will either save or check for equality of the existing silver.
```
TEST_CASE("juice silver", "[file]")
{
    SerializingFactory silver;

    // .. actually use the factory

    CHECK(silver.matches("juice"));
}
```

Diffs=
b329a96f26 Silvers (#9281)

+++ src/math/mat2d.cpp
@@ -204,11 +204,11 @@ TransformComponents Mat2D::decompose() const
     float m0 = m_buffer[0], m1 = m_buffer[1], m2 = m_buffer[2],
           m3 = m_buffer[3];
 
-    float rotation = (float)std::atan2(m1, m0);
+    float rotation = std::atan2(m1, m0);
     float denom = m0 * m0 + m1 * m1;
-    float scaleX = (float)std::sqrt(denom);
+    float scaleX = std::sqrt(denom);
     float scaleY = scaleX == 0.0f ? 0.0f : (m0 * m3 - m2 * m1) / scaleX;
-    float skewX = (float)std::atan2(m0 * m2 + m1 * m3, denom);
+    float skewX = std::atan2(m0 * m2 + m1 * m3, denom);
 
     TransformComponents result;
     result.x(m_buffer[4]);
